OneWire: Check if device present after reset
«The master starts a transmission with a reset pulse, which pulls the
wire to 0 volts for at least 480 µs. This resets every slave device on
the bus. After that, any slave device, if present, shows that it exists
with a "presence" pulse: it holds the bus low for at least 60 µs after
the master releases the bus.»
Source: https://en.wikipedia.org/wiki/1-Wire